+++ CHANGELOG.md
@@ -23,6 +23,9 @@ aliases that have a `Concatenate` special form as their argument.
   `Ellipsis` as an argument. Patch by [Daraan](https://github.com/Daraan).
 - Fix error in subscription of `Unpack` aliases causing nested Unpacks 
   to not be resolved correctly. Patch by [Daraan](https://github.com/Daraan).
+- Backport of CPython PR [#124795](https://github.com/python/cpython/pull/124795)
+  and fix that `TypeAliasType` not raising an error on non-tupple inputs for `type_params`.
+  Patch by [Daraan](https://github.com/Daraan).
 
 # Release 4.12.2 (June 7, 2024)